requirements.txt
python-telegram-bot==22.6
fastapi==0.104.1
uvicorn==0.24.0
